+++ app/src/main/java/com/example/sungwoo/joinactivity/Model/Listview_CoinData.java
@@ -2,14 +2,14 @@ package com.example.sungwoo.joinactivity.Model;
 
 public class Listview_CoinData {
     private String coin_name;
-    private int coin_currentprice;
+    private double coin_currentprice;
     private double coin_change;
     private double coin_transaction_value;
 
     public Listview_CoinData(){
     }
 
-    public Listview_CoinData(String coin_name, int coin_currentprice, double coin_change, double coin_transaction_value) {
+    public Listview_CoinData(String coin_name, double coin_currentprice, double coin_change, double coin_transaction_value) {
         this.coin_name = coin_name;
         this.coin_currentprice = coin_currentprice;
         this.coin_change = coin_change;
@@ -24,11 +24,11 @@ public class Listview_CoinData {
         this.coin_name = coin_name;
     }
 
-    public int getCoin_currentprice() {
+    public double getCoin_currentprice() {
         return coin_currentprice;
     }
 
-    public void setCoin_currentprice(int coin_currentprice) {
+    public void setCoin_currentprice(double coin_currentprice) {
         this.coin_currentprice = coin_currentprice;
     }
